// src/interfaces/life_style_info.interface.ts
export interface LifestyleFormData {
  diet: string;
  smoking: string;
  drinking: string;

  body_type: string;
  physical_status: string;
  fitness_level: string;

  sleep_habit: string;
  wake_up_time: string;

  living_style: string;
  family_type: string;

  social_habits: string;
  travel_habits: string;

  food_habits: string;
  fashion_style: string;

  pet_preference: string;

  driving_habit: string;
  work_life_balance: string;

  religious_lifestyle: string;
}

export const initialLifestyleFormData: LifestyleFormData = {
  diet: "",
  smoking: "",
  drinking: "",

  body_type: "",
  physical_status: "",
  fitness_level: "",

  sleep_habit: "",
  wake_up_time: "",

  living_style: "",
  family_type: "",

  social_habits: "",
  travel_habits: "",

  food_habits: "",
  fashion_style: "",

  pet_preference: "",

  driving_habit: "",
  work_life_balance: "",

  religious_lifestyle: "",
};